+++ src/components/layer-controller/LayerController.js
@@ -5,20 +5,19 @@ import Grid from '@material-ui/core/Grid';
 import Button from '@material-ui/core/Button';
 import AddIcon from '@material-ui/icons/Add';
 import Slider from '@material-ui/core/Slider';
-import InputLabel from '@material-ui/core/InputLabel';
-
 import ExpansionPanel from '@material-ui/core/ExpansionPanel';
-import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
-import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 
 import LayerOptions from './LayerOptions';
 
 import {
-  useExpansionPanelStyles,
-  useExpansionPanelSummaryStyles,
-  useSmallInputLabelStyles,
+  useControllerSectionStyles,
+  StyledExpansionPanelDetails,
+  StyledExpansionPanelSummary,
+  StyledInputLabel,
+  OverflowEllipsisGrid,
 } from './styles';
+
 import { GLOBAL_LABELS } from '../spatial/constants';
 import { getSourceFromLoader, isRgb } from '../../utils';
 import { DOMAINS } from './constants';
@@ -254,28 +253,25 @@ export default function LayerController(props) {
     );
   }
 
-  const classes = useExpansionPanelStyles();
-  const summaryClasses = useExpansionPanelSummaryStyles();
-  const closedOpacityLabelClasses = useSmallInputLabelStyles();
+  const controllerSectionClasses = useControllerSectionStyles();
 
   return (
     <ExpansionPanel
-      className={classes.root}
+      className={controllerSectionClasses.root}
       onChange={(e, expanded) => setIsExpanded(expanded && e?.target?.attributes?.role?.value === 'presentation')}
       TransitionProps={{ enter: false }}
       expanded={isExpanded}
     >
-      <ExpansionPanelSummary
+      <StyledExpansionPanelSummary
         expandIcon={<ExpandMoreIcon />}
         aria-controls={`layer-${name}-controls`}
-        classes={{ ...summaryClasses }}
       >
         <Grid container direction="column" m={1} justify="center">
-          <Grid item>{name}</Grid>
+          <OverflowEllipsisGrid item>{name}</OverflowEllipsisGrid>
           {!isExpanded && (
             <Grid container direction="row" alignItems="center" justify="center">
               <Grid item xs={6}>
-                <InputLabel htmlFor={`layer-${name}-opacity-closed`} classes={closedOpacityLabelClasses}>Opacity:</InputLabel>
+                <StyledInputLabel htmlFor={`layer-${name}-opacity-closed`}>Opacity:</StyledInputLabel>
               </Grid>
               <Grid item xs={6}>
                 <Slider
@@ -293,8 +289,8 @@ export default function LayerController(props) {
             </Grid>
           )}
         </Grid>
-      </ExpansionPanelSummary>
-      <ExpansionPanelDetails className={classes.root}>
+      </StyledExpansionPanelSummary>
+      <StyledExpansionPanelDetails>
         <LayerOptions
           channels={channels}
           labels={labels}
@@ -342,7 +338,7 @@ export default function LayerController(props) {
         >
             Remove Image Layer
         </Button>
-      </ExpansionPanelDetails>
+      </StyledExpansionPanelDetails>
     </ExpansionPanel>
   );
 }